fix(review): apply 10 warning-level review fixes — dead code, misplaced tests, import cleanup, deduplication
