+++ tests/recipe/test_rules_dataflow_capture.py
@@ -9,30 +9,10 @@ import yaml
 
 from autoskillit.core.types import Severity
 from autoskillit.recipe.io import _parse_recipe
-from autoskillit.recipe.schema import Recipe, RecipeStep
 from autoskillit.recipe.validator import run_semantic_rules
-from tests.recipe.conftest import _make_workflow
+from tests.recipe.conftest import _build_merge_worktree_recipe, _make_workflow
 
 pytestmark = [pytest.mark.layer("recipe"), pytest.mark.small]
-
-
-def _build_merge_worktree_recipe(capture: dict) -> Recipe:
-    """Helper: build a minimal Recipe with a merge_worktree step and the given capture dict."""
-    return Recipe(
-        name="test-merge",
-        description="Test merge recipe",
-        summary="merge > done",
-        steps={
-            "merge": RecipeStep(
-                tool="merge_worktree",
-                with_args={"worktree_path": "${{ context.worktree_path }}", "base_branch": "main"},
-                capture=capture,
-                on_success="done",
-                on_failure="done",
-            ),
-            "done": RecipeStep(action="stop", message="Done"),
-        },
-    )
 
 
 # ---------------------------------------------------------------------------
